Fix: unable to parse end of the document marker (...)
In Regexp, '^' and '$' by default match the beginning and end of the
whole string rather than a single line. In Parser.parse(), each line is
a separate string, so this is fine. But in Parser.cleanup(), the whole
document is a single string. To allow '^' and '$' matching the beginning
and end of line in this string, we need modifier 'm' to make RegExp work
in multiline mode.